Topic: 3 Courseworks for Practical Mathematics
Time: 18/19

Introduction: 
The articles are typed in via LaTex. Related data analysis are algorithm are via Rstudio. 
The Coursework 2 of interest is whether the five modes of transport differ in their respective probabilities of different casualty severity. 
The Coursework 3 of interest is fitting a regression model to model house prices (price) and examining the assumption of normal distribution of error.

Data: 
A table for coursework2 is given by my lecture Dr. Peter Foster. 
The data for coursework3 were taken from the public records. They comprise the prices of a random sample of 1728 homes in Saratoga County, New York, USA in 2006, together with the values of 15 other variables which may be influential in explaining the price of a house in this area.